restructure into maven project
