+++ .gitignore
@@ -1,3 +1,6 @@
 *.class
 *.jar
 cesaves/*
+*.iml
+.idea/
+target/
