--- Manifest.txt
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Created-By: 1.7.0_06 (Oracle Corporation)
-Main-Class: Main
