--- build.bat
@@ -1,3 +1,0 @@
-@echo off
-javac *.java
-createjar castleescape.jar Manifest.txt *.class
