--- compile.bat
@@ -1,3 +1,0 @@
-@echo off
-javac *.java
-createjar Manifest.txt castleescape.jar *class
