+++ pom.xml
@@ -1,0 +1,60 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <groupId>com.grahambartley.castleescape</groupId>
+  <artifactId>castle-escape</artifactId>
+  <version>1.0.0</version>
+
+  <properties>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
+  </properties>
+
+  <dependencies>
+    <!-- Add any dependencies here -->
+  </dependencies>
+
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.8.1</version>
+      </plugin>
+      <plugin>
+        <groupId>org.panteleyev</groupId>
+        <artifactId>jpackage-maven-plugin</artifactId>
+        <version>1.6.0</version>
+        <executions>
+          <execution>
+            <id>create-windows-installer</id>
+            <phase>package</phase>
+            <goals>
+              <goal>jpackage</goal>
+            </goals>
+            <configuration>
+              <type>MSI</type>
+              <name>Castle Escape</name>
+              <vendor>Graham Bartley</vendor>
+              <appVersion>${project.version}</appVersion>
+              <icon>${project.basedir}/src/main/resources/castle-escape-logo.ico</icon>
+              <input>target</input>
+              <mainJar>castle-escape-${project.version}.jar</mainJar>
+              <mainClass>com.grahambartley.castleescape.Main</mainClass>
+              <destination>target/dist</destination>
+              <winConsole>true</winConsole>
+              <winDirChooser>true</winDirChooser>
+              <winMenu>true</winMenu>
+              <winMenuGroup>Games</winMenuGroup>
+              <winShortcut>true</winShortcut>
+              <winShortcutPrompt>true</winShortcutPrompt>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+    </plugins>
+  </build>
+</project>
